speedup build - only install nunit if it's not available

+++ build.ps1
@@ -1,2 +1,4 @@
-nuget install nunit -outputdirectory packages
+if(-not(Test-Path "packages\nunit*")) {
+    nuget install nunit -outputdirectory packages
+}
 csc *.cs /r:packages\NUnit.3.10.1\lib\net45\nunit.framework.dll
